Make placeholder accessible
