+++ packages/lexical-playground/src/ui/ContentEditable.css
@@ -21,3 +21,24 @@
     padding-right: 8px;
   }
 }
+
+.ContentEditable__placeholder {
+  font-size: 15px;
+  color: #999;
+  overflow: hidden;
+  position: absolute;
+  text-overflow: ellipsis;
+  top: 8px;
+  left: 8px;
+  right: 8px;
+  user-select: none;
+  white-space: nowrap;
+  display: inline-block;
+  pointer-events: none;
+}
+@media (max-width: 1025px) {
+  .Placeholder__root {
+    left: 8px;
+    right: 8px;
+  }
+}
