+++ packages/lexical-playground/src/ui/ContentEditable.tsx
@@ -11,10 +11,26 @@ import './ContentEditable.css';
 import {ContentEditable} from '@lexical/react/LexicalContentEditable';
 import * as React from 'react';
 
+type Props = {
+  className?: string;
+  placeholderClassName?: string;
+  placeholder: string;
+};
+
 export default function LexicalContentEditable({
   className,
-}: {
-  className?: string;
-}): JSX.Element {
-  return <ContentEditable className={className || 'ContentEditable__root'} />;
+  placeholder,
+  placeholderClassName,
+}: Props): JSX.Element {
+  return (
+    <ContentEditable
+      className={className ?? 'ContentEditable__root'}
+      aria-placeholder={placeholder}
+      placeholder={
+        <div className={placeholderClassName ?? 'ContentEditable__placeholder'}>
+          {placeholder}
+        </div>
+      }
+    />
+  );
 }
