+++ packages/lexical-react/src/LexicalContentEditable.tsx
@@ -6,102 +6,58 @@
  *
  */
 
-import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
-import * as React from 'react';
-import {useCallback, useState} from 'react';
-import useLayoutEffect from 'shared/useLayoutEffect';
+import type {Props as ElementProps} from './shared/LexicalContentEditableElement';
 
-export type Props = {
-  ariaActiveDescendant?: React.AriaAttributes['aria-activedescendant'];
-  ariaAutoComplete?: React.AriaAttributes['aria-autocomplete'];
-  ariaControls?: React.AriaAttributes['aria-controls'];
-  ariaDescribedBy?: React.AriaAttributes['aria-describedby'];
-  ariaExpanded?: React.AriaAttributes['aria-expanded'];
-  ariaLabel?: React.AriaAttributes['aria-label'];
-  ariaLabelledBy?: React.AriaAttributes['aria-labelledby'];
-  ariaMultiline?: React.AriaAttributes['aria-multiline'];
-  ariaOwns?: React.AriaAttributes['aria-owns'];
-  ariaRequired?: React.AriaAttributes['aria-required'];
-  autoCapitalize?: HTMLDivElement['autocapitalize'];
-  'data-testid'?: string | null | undefined;
-} & React.AllHTMLAttributes<HTMLDivElement>;
+import {useLexicalEditable} from '@lexical/react/useLexicalEditable';
 
-export function ContentEditable({
-  ariaActiveDescendant,
-  ariaAutoComplete,
-  ariaControls,
-  ariaDescribedBy,
-  ariaExpanded,
-  ariaLabel,
-  ariaLabelledBy,
-  ariaMultiline,
-  ariaOwns,
-  ariaRequired,
-  autoCapitalize,
-  className,
-  id,
-  role = 'textbox',
-  spellCheck = true,
-  style,
-  tabIndex,
-  'data-testid': testid,
-  ...rest
-}: Props): JSX.Element {
-  const [editor] = useLexicalComposerContext();
-  const [isEditable, setEditable] = useState(false);
+import {useLexicalComposerContext} from './LexicalComposerContext';
+import {ContentEditableElement} from './shared/LexicalContentEditableElement';
+import {useCanShowPlaceholder} from './shared/useCanShowPlaceholder';
 
-  const ref = useCallback(
-    (rootElement: null | HTMLElement) => {
-      // defaultView is required for a root element.
-      // In multi-window setups, the defaultView may not exist at certain points.
-      if (
-        rootElement &&
-        rootElement.ownerDocument &&
-        rootElement.ownerDocument.defaultView
-      ) {
-        editor.setRootElement(rootElement);
-      }
-    },
-    [editor],
-  );
+/* eslint-disable @typescript-eslint/ban-types */
+export type Props = (
+  | {}
+  | {
+      'aria-placeholder': string;
+      placeholder:
+        | ((isEditable: boolean) => null | JSX.Element)
+        | null
+        | JSX.Element;
+    }
+) &
+  ElementProps;
+/* eslint-enable @typescript-eslint/ban-types */
 
-  useLayoutEffect(() => {
-    setEditable(editor.isEditable());
-    return editor.registerEditableListener((currentIsEditable) => {
-      setEditable(currentIsEditable);
-    });
-  }, [editor]);
+export function ContentEditable(props: Props): JSX.Element {
+  let placeholder = null;
+  if ('placeholder' in props) {
+    placeholder = props.placeholder;
+  }
 
   return (
-    <div
-      {...rest}
-      aria-activedescendant={!isEditable ? undefined : ariaActiveDescendant}
-      aria-autocomplete={!isEditable ? 'none' : ariaAutoComplete}
-      aria-controls={!isEditable ? undefined : ariaControls}
-      aria-describedby={ariaDescribedBy}
-      aria-expanded={
-        !isEditable
-          ? undefined
-          : role === 'combobox'
-          ? !!ariaExpanded
-          : undefined
-      }
-      aria-label={ariaLabel}
-      aria-labelledby={ariaLabelledBy}
-      aria-multiline={ariaMultiline}
-      aria-owns={!isEditable ? undefined : ariaOwns}
-      aria-readonly={!isEditable ? true : undefined}
-      aria-required={ariaRequired}
-      autoCapitalize={autoCapitalize}
-      className={className}
-      contentEditable={isEditable}
-      data-testid={testid}
-      id={id}
-      ref={ref}
-      role={role}
-      spellCheck={spellCheck}
-      style={style}
-      tabIndex={tabIndex}
-    />
+    <>
+      <ContentEditableElement {...props} />
+      <Placeholder content={placeholder} />
+    </>
   );
+}
+
+function Placeholder({
+  content,
+}: {
+  content: ((isEditable: boolean) => null | JSX.Element) | null | JSX.Element;
+}): null | JSX.Element {
+  const [editor] = useLexicalComposerContext();
+  const showPlaceholder = useCanShowPlaceholder(editor);
+  const editable = useLexicalEditable();
+
+  if (!showPlaceholder) {
+    return null;
+  }
+
+  if (typeof content === 'function') {
+    return content(editable);
+  } else {
+    return content;
+  }
 }
