+++ packages/lexical-react/src/LexicalPlainTextPlugin.tsx
@@ -16,11 +16,12 @@ import {usePlainTextSetup} from './shared/usePlainTextSetup';
 
 export function PlainTextPlugin({
   contentEditable,
-  placeholder,
+  // TODO Remove. This property is now part of ContentEditable
+  placeholder = null,
   ErrorBoundary,
 }: {
   contentEditable: JSX.Element;
-  placeholder:
+  placeholder?:
     | ((isEditable: boolean) => null | JSX.Element)
     | null
     | JSX.Element;
@@ -39,6 +40,7 @@ export function PlainTextPlugin({
   );
 }
 
+// TODO Remove
 function Placeholder({
   content,
 }: {
